requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
boto3==1.26.165
botocore==1.29.165
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.3
googlemaps==4.10.0
idna==3.4
jmespath==1.0.1
msgpack==1.0.5
oauthlib==3.2.2
Pillow==10.0.0
pipdeptree==2.9.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
